Enhance documentation and refactor project structure
- Added an image to the main Readme to improve visual appeal.
- Updated .gitignore to exclude .DS_Store files, ensuring a cleaner
repository state on macOS.
- Major overhaul of the Series #01 ReadMe, replacing placeholder code
with an extensive, informative introduction and examples of Node.js
concepts.
- Removed touch.ts example script and reintroduced it under the
src/examples directory, thereby centralizing example code.
- Updated package name and license in package-lock.json, aligning with
recent changes.
- Revised package.json to improve project metadata, including author
info and project repository details, aiding in project discoverability.
- Introduced cat.ts example to showcase file content reading in Node.js.
- Crafted a new src/main.ts script that demonstrates basic command-line
argument handling.
- Modified tsconfig.json to exclude node_modules explicitly and updated
the target to "ES2017" for consistent capitalization.
- Slight content adjustment to file.txt, ensuring its relevance to the
updated project.
- Pin dependencies to match the current LTS version of Node.js.

By reorganizing example scripts and expanding documentation, users now
have a clearer, more structured learning path. Additionally, the project
is better equipped for contributions thanks to updated metadata and
cleaner repository management.

Refs #series-01
