+++ series-01/.gitignore
@@ -1,1 +1,2 @@
 node_modules
+.DS_Store
