+++ series-01/file.txt
@@ -1,1 +1,1 @@
-dummy content for to cat command
+this is dummy file content
